+++ resources/views/components/input/controls/flatpickr.blade.php
@@ -14,6 +14,7 @@
         })"
         {{ $attributes->class([
                 'bdr-red z-100 placeholder-red-400' => $errors->has($for),
+                'w-full'
             ])->except(['for']) }}
         type="text">